# services/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Nook API",
    "description": "Collaborative workspace API. The document is generated from the same Zod schemas the server validates with, so it cannot drift from the implementation.",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/v1/pages/{pageId}/blocks": {
      "get": {
        "tags": [
          "blocks"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "pageId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "pageId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "parentId": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                      },
                      "type": {
                        "type": "string"
                      },
                      "content": {
                        "type": "object",
                        "additionalProperties": {}
                      },
                      "position": {
                        "type": "number"
                      }
                    },
                    "required": [
                      "id",
                      "pageId",
                      "parentId",
                      "type",
                      "content",
                      "position"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "blocks"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "parentId": {
                    "type": "string",
                    "format": "uuid",
                    "nullable": true
                  },
                  "type": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 64
                  },
                  "content": {
                    "type": "object",
                    "additionalProperties": {},
                    "default": {}
                  },
                  "position": {
                    "type": "number",
                    "default": 0
                  }
                },
                "required": [
                  "type"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "pageId",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "pageId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "parentId": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    },
                    "type": {
                      "type": "string"
                    },
                    "content": {
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "position": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "id",
                    "pageId",
                    "parentId",
                    "type",
                    "content",
                    "position"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/workspaces/{workspaceId}/search": {
      "get": {
        "summary": "Full-text search across a workspace's blocks",
        "tags": [
          "blocks"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200
            },
            "in": "query",
            "name": "q",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "workspaceId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "blockId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "pageId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "pageTitle": {
                        "type": "string"
                      },
                      "snippet": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "blockId",
                      "pageId",
                      "pageTitle",
                      "snippet"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}